backend/pyproject.toml
[tool.poetry]
name = "polydash"
version = "0.1.0"
description = ""
authors = ["V.G. Bulavintsev <golem.md@gmail.com>", "super-dev-rust <n.tpronin@gmail.com>", "ShockedShodan <shockedshodan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pony = "^0.7.16"
uvicorn = "^0.22.0"
fastapi = "^0.98.0"
requests = "^2.31.0"
pynng = "^0.7.2"
psycopg2-binary = "^2.9.7"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

